+++ NEWS.md
@@ -1,8 +1,16 @@
+# pestr 0.8.0
+
+## Function updates
+
+* `eppo_tabletool` functions now can use "raw" eppocodes instead of result of `eppo_names_tables` function result to query **EPPO Data Services** and **EPPO Global Database**. New functionality works with `raw_eppocodes` and 'use_raw_codes` parameters.
+
 # pestr 0.7.4
 
 ## New and updated test.
 
+* updated sample database (on 2020-Jan-06).
 * updated `eppo_tabletools_hosts` test including changes in function behaviour.
+* if `eppo_token` is incorrect, functions return invisible `NULL`
 
 ## Fixed issues and smaller functionalities